+++ frontend/src/components/FleetView/types.ts
@@ -1,0 +1,10 @@
+export interface NodeUpdateStatus {
+    nodeId: number;
+    name: string;
+    type: 'local' | 'remote';
+    version: string | null;
+    latestVersion: string | null;
+    updateAvailable: boolean;
+    updateStatus: 'updating' | 'completed' | 'timeout' | 'failed' | null;
+    error?: string | null;
+}
